Added dependency enforcement.
When Data<T>() is invoked, it now checks to make sure that
T is either a dependency that has been previously declared,
or is the current system.  (Otherwise it asserts.)

Also changed the storage containers for the dependency lists
to be unordered_set, since that fits the usage case slightly
better.